Bump to revision 237 and fix migration systems
Update rsprot dependencies, Build revision, cache install handling, and rev237.5-derived symbols.

Touch cache decoders/enrichers, config refs, protocol handlers, zone/map packets, login setup, IF script triggers, music/audio settings, Lumbridge music coverage, docs, and license branding to resolve migration issues after the revision bump.

+++ api/cache/src/main/kotlin/org/rsmod/api/cache/types/comp/ComponentTypeEncoder.kt
@@ -90,8 +90,8 @@ public object ComponentTypeEncoder {
             }
 
             if (type == 6) {
-                data.writeShort(model)
-                data.writeShort(secondaryModel)
+                data.writeInt(model)
+                data.writeInt(secondaryModel)
                 data.writeShort(modelAnim)
                 data.writeShort(secondaryModelAnim)
                 data.writeShort(modelZoom)
@@ -149,7 +149,7 @@ public object ComponentTypeEncoder {
             }
 
             if (type == 6) {
-                data.writeShort(model)
+                data.writeInt(model)
                 data.writeShort(modelX)
                 data.writeShort(modelY)
                 data.writeShort(modelAngleX)
